refector authentication logic
