+++ Gemfile
@@ -19,6 +19,7 @@ gem 'puma', '~> 3.7'
 # gem 'redis', '~> 4.0'
 # Use ActiveModel has_secure_password
 gem 'bcrypt', '~> 3.1.7'
+gem 'pry'
 
 # Use Capistrano for deployment
 # gem 'capistrano-rails', group: :development
@@ -28,6 +29,7 @@ gem 'bcrypt', '~> 3.1.7'
 group :production do
   gem 'pg'
   gem 'rails_12factor'
+  gem 'pry'
 end
 
 group :development, :test do
