+++ app/controllers/authentication_controller.rb
@@ -1,30 +1,34 @@
 class AuthenticationController < ApplicationController
   skip_before_action :authenticate, only: [:create]
 
-  def initialize
-    @auth = Auth::Authentication.new
-  end
 
   def create
     @user = User.find_by(email: auth_params[:email])
     if @user
       @user.authenticate(auth_params[:password])
-      jwt = @auth.issue({ user: @user.id })
+      jwt = issue({ user: @user.id })
       @user.update({ login_status: true })
       render json: { user: @user, jwt: jwt, message: 'You\'re in!' }
     else
-      render json: { message: 'Invalid login details' }, status: :bad_request
+      render json: { message: 'Invalid login details', error: @user.errors.full_messages }, status: :bad_request
     end
   end
 
   def destroy
+    # binding.pry
+    puts "###################"
+    puts @current_user
+    puts "###################"
+    # binding.pry
+    puts "###################"
+
     if @current_user
       user = set_user
       user.update({ login_status: false })
       @current_user = nil
       render json: { message: 'Thanks for using our app, good bye' }, status: :ok
     else
-      render json: { message: 'You have to be logged in' }, status: :unauthorized
+      render json: { message: 'You have to be logged in authentication',  error: @current_user.errors.full_messages }, status: :unauthorized
     end
   end
